pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "umbc-campus-knowledge-assistant"
version = "0.1.0"
description = "AI-powered campus knowledge assistant using UMBC institutional data"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Soumith" }]
dependencies = [
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "spacy>=3.7.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "jinja2>=3.1.0",
  "pyspellchecker>=0.8.1",
  "rapidfuzz>=3.9.0",
]

[project.optional-dependencies]
llm = ["openai>=1.40.0"]
dense = ["sentence-transformers>=3.0.0", "torch>=2.3.0"]
dev = ["pytest>=8.0.0", "ruff>=0.6.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
